Move FlavorConfigBuilder from testutils to make it osgi visible
